Added Janky Concept Art for Model #JK.D45.NK
Super-basic drawing of snake-based mechanical 'rassler.  Based off of Jake "The Snake" Roberts, but renamed with a robot model number to avoid litigation.  Finisher can be the DDT.  Would be cool to hit it onto a saw blade trap or explosive device.  I'm leaning towards WITH legs and either a shortened tail or no tail; just give the overall design a snake-like look with a mechanical edge.  Of course we can keep the tail and add rollers onto the bottom.